games/gcompris-qt: import gcompris-qt-3.1
GCompris is a high quality educational software suite, including
a large number of activities for children aged 2 to 10. Some of
the activities are game orientated, but still educational.

Currently GCompris offers more than 100 activities, and more are
being developed. GCompris is free software, it means that you can
adapt it to your own needs, improve it and, most importantly, share
it with children everywhere.